+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- api, api-contract, types 16.3.1
 - hw-ledger, keyring, util, util-crypto 13.5.3
 - extension-dapp 0.61.1
 - extension-dapp 0.60.1